feat: move enum folder to out of app folder
