+++ src/app/dashboard/products/delete-form.tsx
@@ -2,7 +2,7 @@ import { FormEvent, useState } from 'react'
 import { toast } from 'react-toastify'
 import { confirmAlert } from 'react-confirm-alert'
 
-import { Types } from '@/app/enum/types'
+import { Types } from '@/enum/types'
 import DialogConfirm from '@/app/ui/dashboard/dialog-confirm/dialog-confirm'
 import { DeleteButton } from '../../ui/dashboard/delete-button/delete-button'
 import { deleteProductServerAction } from '@/app/actions/products/delete-product-action'
