+++ src/app/ui/dashboard/mobile-sidebar/mobile-sidebar-client.tsx
@@ -3,7 +3,7 @@
 import Image from 'next/image'
 import { MdClose } from 'react-icons/md'
 
-import { TypesColors } from '@/app/enum/typeColors'
+import { TypesColors } from '@/enum/typeColors'
 import { menuItemsData } from '@/app/utils/menuItemsData'
 import MenuLink from '@/app/ui/dashboard/menuLink/menulink'
 import { useSidebarMobileStore } from '@/store/sidebarMobileStore'
