+++ src/app/ui/dashboard/sidebar/sidebar.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image'
 
 import { auth } from '@/app/auth'
-import { TypesColors } from '@/app/enum/typeColors'
+import { TypesColors } from '@/enum/typeColors'
 import { menuItemsData } from '@/app/utils/menuItemsData'
 import MenuLink from '@/app/ui/dashboard/menuLink/menulink'
 import LogoutButton from '@/app/ui/dashboard/logout-button/logout-button'
